refactor: move FidelityBondAdvanced component to own file
